✨ Update fileServer Code from Dev branch
- Update fileServer Code from Dev branch

+++ src/main/java/com/myhome/server/api/service/FileServerPrivateServiceImpl.java
@@ -3,6 +3,7 @@ package com.myhome.server.api.service;
 import com.myhome.server.api.dto.FileServerPrivateDto;
 import com.myhome.server.api.dto.FileServerPrivateTrashDto;
 import com.myhome.server.api.dto.FileServerPublicDto;
+import com.myhome.server.api.dto.FileServerPublicTrashDto;
 import com.myhome.server.config.jwt.JwtTokenProvider;
 import com.myhome.server.db.entity.*;
 import com.myhome.server.db.repository.FileServerPrivateRepository;
@@ -81,6 +82,12 @@ public class FileServerPrivateServiceImpl implements FileServerPrivateService {
     }
 
     @Override
+    public List<FileServerPrivateTrashEntity> findTrashAll() {
+        List<FileServerPrivateTrashEntity> list = trashRepository.findAll();
+        return list;
+    }
+
+    @Override
     public HttpHeaders getHttpHeaders(Path path, String fileName) throws IOException {
         String contentType = Files.probeContentType(path); // content type setting
 
@@ -138,6 +145,33 @@ public class FileServerPrivateServiceImpl implements FileServerPrivateService {
     }
 
     @Override
+    public void mkdir(String path, String token) {
+        File file = new File(path);
+        file.mkdir();
+//        String[] paths = path.split(File.separator);
+        String[] paths = path.split("\\\\");
+        String name = paths[paths.length-1];
+        StringBuilder location = new StringBuilder();
+        for(int i=0;i<paths.length-1;i++){
+            location.append(paths[i]).append(File.separator);
+        }
+        String pk = jwtTokenProvider.getUserPk(token);
+        Optional<UserEntity> userEntity = service.findById(pk);
+
+        FileServerPrivateDto dto = new FileServerPrivateDto(
+                path, // file path (need to change)
+                name, // file name
+                UUID.randomUUID().toString(), // file name to change UUID
+                "dir",
+                0, // file size(KB)
+                userEntity.get().getName(),
+                location.toString(), // file folder path (need to change)
+                0
+        );
+        repository.save(new FileServerPrivateEntity(dto));
+    }
+
+    @Override
     public boolean existsByPath(String path) {
         boolean result = repository.existsByPath(path);
         return result;
@@ -149,24 +183,115 @@ public class FileServerPrivateServiceImpl implements FileServerPrivateService {
         if(ObjectUtils.isEmpty(entity)){
             return -1;
         }
-        String testInPath = "C:\\Users\\SonJunHyeok\\Desktop\\a.txt"; // test filePath
-        String filePath = testInPath;
-        //        String filePath = entity.getPath();
+//        String testInPath = "C:\\Users\\SonJunHyeok\\Desktop\\a.txt"; // test filePath
+//        String filePath = testInPath;
+//        String filePath = entity.getPath();
+//        File file = new File(filePath);
 
-        File file = new File(filePath);
+        File file = new File(path);
 
         if(file.exists()){ // check file exist
-            if(file.delete()){ // if file exist, delete file
-                System.out.println("delete success");
+            if(file.isDirectory()){
+                deleteFolder(file.getPath());
             }
             else{
-                System.out.println("delete failed");
+                if(file.delete()){ // if file exist, delete file
+                    System.out.println("delete success");
+                    repository.deleteByPath(path);
+                }
+                else{
+                    System.out.println("delete failed");
+                }
+            }
+
+        }
+        return 0;
+    }
+    private void moveFolder(String path, String destPath){
+        File dir = new File(path);
+        File[] files = dir.listFiles();
+        if(files != null){
+            for(File file : files){
+                if(file.isDirectory()) {
+                    moveFolder(path+File.separator+file.getName(), destPath+File.separator+file.getName());
+                    String dirName = file.getName();
+                    FileServerPrivateEntity entity = repository.findByPath(path+File.separator+dirName);
+                    if(entity != null){
+                        FileServerPrivateTrashDto trashDto = new FileServerPrivateTrashDto(entity);
+                        FileServerPrivateTrashEntity trashEntity = new FileServerPrivateTrashEntity(trashDto);
+                        trashRepository.save(trashEntity);
+                    }
+                    else{
+                        FileServerPrivateTrashDto trashDto = new FileServerPrivateTrashDto();
+                        trashDto.setPath(path+File.separator+file.getName());
+                        trashDto.setLocation(path);
+                        trashDto.setName(file.getName());
+                        trashDto.setSize(0);
+                        trashDto.setState(0);
+                        trashDto.setType("dir");
+                        trashDto.setUuidName(UUID.randomUUID().toString());
+                        trashRepository.save(new FileServerPrivateTrashEntity(trashDto));
+                    }
+                    if(file.delete()){
+                        System.out.println("FileServerPublicService Folder delete Success : "+file.getPath());
+                    }
+                    else{
+                        System.out.println("FileServerPublicService Folder delete Fail : "+file.getPath());
+                    }
+                }
+                else {
+                    try{
+                        FileServerPrivateEntity entity = repository.findByPath(path);
+                        File in = new File(entity.getPath());
+                        File out = new File(destPath);
+                        FileCopyUtils.copy(in, out); // copy file from origin location to new location
+                        if(in.exists()){ // check origin file exist
+                            if(in.delete()){ // if file exist
+                                System.out.println("delete success");
+                                repository.deleteByPath(entity.getPath());
+                            }
+                            else{
+                                System.out.println("delete failed");
+                            }
+                        }
+                    }
+                    catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
             }
         }
-        long result = repository.deleteByPath(path); // delete file info from DB
-        return result;
     }
-
+    private void deleteFolder(String path){
+        File dir = new File(path);
+        File[] files = dir.listFiles();
+        FileServerPrivateTrashEntity entity;
+        if(files != null){
+            for(File file : files){
+                if(file.isDirectory()){
+                    deleteFolder(file.getPath());
+                    if(file.delete()){
+                        entity = trashRepository.findByPath(path);
+                        if(entity != null) trashRepository.delete(trashRepository.findByPath(path));
+                        System.out.println("FileServerPrivateTrashEntity delete folder Success");
+                    }
+                    else{
+                        System.out.println("FileServerPrivateTrashEntity delete folder failed");
+                    }
+                }
+                else{
+                    if(file.delete()){
+                        entity = trashRepository.findByPath(path);
+                        if(entity != null) trashRepository.delete(trashRepository.findByPath(path));
+                        System.out.println("FileServerPrivateTrashEntity delete file Success");
+                    }
+                    else{
+                        System.out.println("FileServerPrivateTrashEntity delete file failed");
+                    }
+                }
+            }
+        }
+    }
     @Override
     public int moveFile(String uuid, String location) {
         FileServerPrivateEntity entity = repository.findByUuid(uuid);
@@ -205,30 +330,33 @@ public class FileServerPrivateServiceImpl implements FileServerPrivateService {
     @Override
     public int moveTrash(String uuid) {
         FileServerPrivateEntity entity = repository.findByUuid(uuid);
-        if(entity != null){
+        if(entity != null) {
             FileServerPrivateTrashDto dto = new FileServerPrivateTrashDto(entity);
             FileServerPrivateTrashEntity trashEntity = new FileServerPrivateTrashEntity(dto);
             trashRepository.save(trashEntity);
+            String trashPath = diskPath + "trash" + File.separator + dto.getName();
 
-            String trashPath = diskPath+"trash"+File.separator+dto.getName();
-            try{
-                File in = new File(entity.getPath());
-                File out = new File(trashPath);
-                FileCopyUtils.copy(in, out); // copy file from origin location to new location
-                if(in.exists()){ // check origin file exist
-                    if(in.delete()){ // if file exist
-                        System.out.println("delete success");
-                        repository.deleteByPath(entity.getPath());
-                        return 0;
-                    }
-                    else{
-                        System.out.println("delete failed");
-                    }
-                }
+            if (entity.getType().equals("dir")) {
+                moveFolder(entity.getPath(), trashPath);
             }
-            catch (IOException e) {
-                e.printStackTrace();
-                return -1;
+            else {
+                try {
+                    File in = new File(entity.getPath());
+                    File out = new File(trashPath);
+                    FileCopyUtils.copy(in, out); // copy file from origin location to new location
+                    if (in.exists()) { // check origin file exist
+                        if (in.delete()) { // if file exist
+                            System.out.println("delete success");
+                            repository.deleteByPath(entity.getPath());
+                            return 0;
+                        } else {
+                            System.out.println("delete failed");
+                        }
+                    }
+                } catch (IOException e) {
+                    e.printStackTrace();
+                    return -1;
+                }
             }
         }
         return -1;
